Fix: Implementacion de tipo size_t

+++ A_AR/main.cpp
@@ -5,7 +5,7 @@ using namespace std;
 
 bool factible(const vector<int> &S, int x, const vector<int> &c){
     int c_act = c[x];
-    for(int i = 0; i < S.size(); i++){
+    for(size_t i = 0; i < S.size(); i++){
         if(S[i] == x){
             c_act -= 1;
         }
@@ -23,7 +23,7 @@ void insertar(vector<int> &S, int x, const vector<vector<int> > &b, int t){
 
 int objetivo(const vector<int> &S, const vector<vector<int> > &b){
     int bt = 0;
-    for(int i = 0; i < S.size(); i++){
+    for(size_t i = 0; i < S.size(); i++){
         if(S[i] == -1){
             bt = 0;
             return bt;
@@ -36,7 +36,7 @@ int objetivo(const vector<int> &S, const vector<vector<int> > &b){
 int seleccionar(const vector<int> &C, int j, const vector<vector<int> > &b){
     int candidato = C[0];
     int max_b = b[C[0]][j];
-    for(int i = 0; i < C.size(); i++){
+    for(size_t i = 0; i < C.size(); i++){
         if(max_b < b[C[i]][j]){
             max_b = b[C[i]][j];
             candidato = C[i];
@@ -47,7 +47,7 @@ int seleccionar(const vector<int> &C, int j, const vector<vector<int> > &b){
 }
 
 bool solucion(const vector<int> &S, const vector<int> &c, const vector<vector<int> > &b, int nt){
-    if(S.size() != nt){ 
+    if(S.size() != (size_t)nt){ 
         return false; 
     }
     vector<int> c_copia = c;
@@ -66,7 +66,7 @@ bool solucion(const vector<int> &S, const vector<int> &c, const vector<vector<in
 
 void actualizar_cjto(vector<int> &C, const vector<int> &S, const vector<int> &c, int x){
     int cont = 0;
-    for(int i = 0; i < S.size(); i++){
+    for(size_t i = 0; i < S.size(); i++){
         if(S[i] == x){
             cont += 1;
         }
@@ -129,7 +129,7 @@ int main(void){
         vector<int> S = avance_rapido(C, b, c, nw, nt);
 
         cout << objetivo(S, b) << endl;
-        for(int j = 0; j < S.size(); j++){
+        for(size_t j = 0; j < S.size(); j++){
             cout << S[j] << " "; 
         }
         cout << endl;
